requirements.txt
Flask==2.3.3
PyMySQL==1.1.0
python-dotenv==1.0.0
Werkzeug==2.3.7